Struct google_dataproc1::HadoopJob

pub struct HadoopJob {
    pub jar_file_uris: Option<Vec<String>>,
    pub logging_config: Option<LoggingConfig>,
    pub args: Option<Vec<String>>,
    pub file_uris: Option<Vec<String>>,
    pub main_class: Option<String>,
    pub archive_uris: Option<Vec<String>>,
    pub main_jar_file_uri: Option<String>,
    pub properties: Option<HashMap<String, String>>,
}

A Cloud Dataproc job for running Hadoop MapReduce jobs on YARN.

This type is not used in any activity, and only used as part of another schema.

Fields

jar_file_uris

[Optional] Jar file URIs to add to the CLASSPATHs of the Hadoop driver and tasks.

logging_config

[Optional] The runtime log config for job execution.

args

[Optional] The arguments to pass to the driver. Do not include arguments, such as -libjars or -Dfoo=bar, that can be set as job properties, since a collision may occur that causes an incorrect job submission.

file_uris

[Optional] HCFS (Hadoop Compatible Filesystem) URIs of files to be copied to the working directory of Hadoop drivers and distributed tasks. Useful for naively parallel tasks.

main_class

The name of the driver's main class. The jar file containing the class must be in the default CLASSPATH or specified in jar_file_uris.

archive_uris

[Optional] HCFS URIs of archives to be extracted in the working directory of Hadoop drivers and tasks. Supported file types: .jar, .tar, .tar.gz, .tgz, or .zip.

main_jar_file_uri

The HCFS URI of the jar file containing the main class. Examples: 'gs://foo-bucket/analytics-binaries/extract-useful-metrics-mr.jar' 'hdfs:/tmp/test-samples/custom-wordcount.jar' 'file:///home/usr/lib/hadoop-mapreduce/hadoop-mapreduce-examples.jar'

properties

[Optional] A mapping of property names to values, used to configure Hadoop. Properties that conflict with values set by the Cloud Dataproc API may be overwritten. Can include properties set in /etc/hadoop/conf/*-site and classes in user code.