Struct google_dataproc1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Dataproc hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dataproc1 as dataproc1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dataproc1::Dataproc;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dataproc::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `regions_clusters_create(...)`, `regions_clusters_delete(...)`, `regions_clusters_diagnose(...)`, `regions_clusters_get(...)`, `regions_clusters_list(...)`, `regions_clusters_patch(...)`, `regions_jobs_cancel(...)`, `regions_jobs_delete(...)`, `regions_jobs_get(...)`, `regions_jobs_list(...)`, `regions_jobs_submit(...)`, `regions_operations_cancel(...)`, `regions_operations_delete(...)`, `regions_operations_get(...)` and `regions_operations_list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn regions_clusters_get(&self, project_id: &str, region: &str, cluster_name: &str) -> ProjectRegionClusterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the resource representation for a cluster in a project.

Arguments

projectId - [Required] The ID of the Google Cloud Platform project that the cluster belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

clusterName - [Required] The cluster name.

fn regions_clusters_list(&self, project_id: &str, region: &str) -> ProjectRegionClusterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all regions/{region}/clusters in a project.

Arguments

projectId - [Required] The ID of the Google Cloud Platform project that the cluster belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

fn regions_jobs_list(&self, project_id: &str, region: &str) -> ProjectRegionJobListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists regions/{region}/jobs in a project.

Arguments

projectId - [Required] The ID of the Google Cloud Platform project that the job belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

fn regions_clusters_patch(&self, request: Cluster, project_id: &str, region: &str, cluster_name: &str) -> ProjectRegionClusterPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a cluster in a project.

Arguments

request - No description provided.

projectId - [Required] The ID of the Google Cloud Platform project the cluster belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

clusterName - [Required] The cluster name.

fn regions_jobs_get(&self, project_id: &str, region: &str, job_id: &str) -> ProjectRegionJobGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the resource representation for a job in a project.

Arguments

projectId - [Required] The ID of the Google Cloud Platform project that the job belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

jobId - [Required] The job ID.

fn regions_operations_cancel(&self, name: &str) -> ProjectRegionOperationCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation.

Arguments

name - The name of the operation resource to be cancelled.

fn regions_operations_get(&self, name: &str) -> ProjectRegionOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

fn regions_clusters_delete(&self, project_id: &str, region: &str, cluster_name: &str) -> ProjectRegionClusterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a cluster in a project.

Arguments

projectId - [Required] The ID of the Google Cloud Platform project that the cluster belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

clusterName - [Required] The cluster name.

fn regions_clusters_diagnose(&self, request: DiagnoseClusterRequest, project_id: &str, region: &str, cluster_name: &str) -> ProjectRegionClusterDiagnoseCall<'a, C, A>

Create a builder to help you perform the following task:

Gets cluster diagnostic information. After the operation completes, the Operation.response field contains DiagnoseClusterOutputLocation.

Arguments

request - No description provided.

projectId - [Required] The ID of the Google Cloud Platform project that the cluster belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

clusterName - [Required] The cluster name.

fn regions_jobs_submit(&self, request: SubmitJobRequest, project_id: &str, region: &str) -> ProjectRegionJobSubmitCall<'a, C, A>

Create a builder to help you perform the following task:

Submits a job to a cluster.

Arguments

request - No description provided.

projectId - [Required] The ID of the Google Cloud Platform project that the job belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

fn regions_clusters_create(&self, request: Cluster, project_id: &str, region: &str) -> ProjectRegionClusterCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a cluster in a project.

Arguments

request - No description provided.

projectId - [Required] The ID of the Google Cloud Platform project that the cluster belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

fn regions_jobs_delete(&self, project_id: &str, region: &str, job_id: &str) -> ProjectRegionJobDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the job from the project. If the job is active, the delete fails, and the response returns FAILED_PRECONDITION.

Arguments

projectId - [Required] The ID of the Google Cloud Platform project that the job belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

jobId - [Required] The job ID.

fn regions_operations_delete(&self, name: &str) -> ProjectRegionOperationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

fn regions_operations_list(&self, name: &str) -> ProjectRegionOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding below allows API services to override the binding to use different resource name schemes, such as users/*/operations.

Arguments

name - The name of the operation collection.

fn regions_jobs_cancel(&self, request: CancelJobRequest, project_id: &str, region: &str, job_id: &str) -> ProjectRegionJobCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts a job cancellation request. To access the job resource after cancellation, call regions/{region}/jobs.list or regions/{region}/jobs.get.

Arguments

request - No description provided.

projectId - [Required] The ID of the Google Cloud Platform project that the job belongs to.

region - [Required] The Cloud Dataproc region in which to handle the request.

jobId - [Required] The job ID.