Struct google_dataproc1::ClusterConfig

pub struct ClusterConfig {
    pub software_config: Option<SoftwareConfig>,
    pub config_bucket: Option<String>,
    pub gce_cluster_config: Option<GceClusterConfig>,
    pub initialization_actions: Option<Vec<NodeInitializationAction>>,
    pub worker_config: Option<InstanceGroupConfig>,
    pub secondary_worker_config: Option<InstanceGroupConfig>,
    pub master_config: Option<InstanceGroupConfig>,
}

The cluster config.

This type is not used in any activity, and only used as part of another schema.

Fields

software_config

[Optional] The config settings for software inside the cluster.

config_bucket

[Optional] A Google Cloud Storage staging bucket used for sharing generated SSH keys and config. If you do not specify a staging bucket, Cloud Dataproc will determine an appropriate Cloud Storage location (US, ASIA, or EU) for your cluster's staging bucket according to the Google Compute Engine zone where your cluster is deployed, and then it will create and manage this project-level, per-location bucket for you.

gce_cluster_config

[Optional] The shared Google Compute Engine config settings for all instances in a cluster.

initialization_actions

[Optional] Commands to execute on each node after config is completed. By default, executables are run on master and all worker nodes. You can test a node's role metadata to run an executable on a master or worker node, as shown below: ROLE=$(/usr/share/google/get_metadata_value attributes/role) if [[ "${ROLE}" == 'Master' ]]; then ... master specific actions ... else ... worker specific actions ... fi

worker_config

[Optional] The Google Compute Engine config settings for worker instances in a cluster.

secondary_worker_config

[Optional] The Google Compute Engine config settings for additional worker instances in a cluster.

master_config

[Optional] The Google Compute Engine config settings for the master instance in a cluster.