Struct google_dataproc1::GceClusterConfig

pub struct GceClusterConfig {
    pub network_uri: Option<String>,
    pub subnetwork_uri: Option<String>,
    pub service_account_scopes: Option<Vec<String>>,
    pub tags: Option<Vec<String>>,
    pub metadata: Option<HashMap<String, String>>,
    pub zone_uri: Option<String>,
}

Common config settings for resources of Google Compute Engine cluster instances, applicable to all instances in the cluster.

This type is not used in any activity, and only used as part of another schema.

Fields

network_uri

The Google Compute Engine network to be used for machine communications. Cannot be specified with subnetwork_uri. If neither network_uri nor subnetwork_uri is specified, the "default" network of the project is used, if it exists. Cannot be a "Custom Subnet Network" (see https://cloud.google.com/compute/docs/subnetworks for more information). Example: https://www.googleapis.com/compute/v1/projects/[project_id]/regions/global/default.

subnetwork_uri

The Google Compute Engine subnetwork to be used for machine communications. Cannot be specified with network_uri. Example: https://www.googleapis.com/compute/v1/projects/[project_id]/regions/us-east1/sub0.

service_account_scopes

The URIs of service account scopes to be included in Google Compute Engine instances. The following base set of scopes is always included: * https://www.googleapis.com/auth/cloud.useraccounts.readonly * https://www.googleapis.com/auth/devstorage.read_write * https://www.googleapis.com/auth/logging.write If no scopes are specfied, the following defaults are also provided: * https://www.googleapis.com/auth/bigquery * https://www.googleapis.com/auth/bigtable.admin.table * https://www.googleapis.com/auth/bigtable.data * https://www.googleapis.com/auth/devstorage.full_control

tags

The Google Compute Engine tags to add to all instances.

metadata

The Google Compute Engine metadata entries to add to all instances.

zone_uri

[Required] The zone where the Google Compute Engine cluster will be located. Example: https://www.googleapis.com/compute/v1/projects/[project_id]/zones/[zone].