Struct google_dataproc1::HiveJob

pub struct HiveJob {
    pub query_file_uri: Option<String>,
    pub script_variables: Option<HashMap<String, String>>,
    pub continue_on_failure: Option<bool>,
    pub jar_file_uris: Option<Vec<String>>,
    pub query_list: Option<QueryList>,
    pub properties: Option<HashMap<String, String>>,
}

A Cloud Dataproc job for running Hive queries on YARN.

This type is not used in any activity, and only used as part of another schema.

Fields
query_file_uri: Option<String>

The HCFS URI of the script that contains Hive queries.

script_variables: Option<HashMap<String, String>>

[Optional] Mapping of query variable names to values (equivalent to the Hive command: SET name="value";).

continue_on_failure: Option<bool>

[Optional] Whether to continue executing queries if a query fails. The default value is false. Setting to true can be useful when executing independent parallel queries.

jar_file_uris: Option<Vec<String>>

[Optional] HCFS URIs of jar files to add to the CLASSPATH of the Hive server and Hadoop MapReduce (MR) tasks. Can contain Hive SerDes and UDFs.

query_list: Option<QueryList>

A list of queries.

properties: Option<HashMap<String, String>>

[Optional] A mapping of property names and values, used to configure Hive. Properties that conflict with values set by the Cloud Dataproc API may be overwritten. Can include properties set in /etc/hadoop/conf/*-site.xml, /etc/hive/conf/hive-site.xml, and classes in user code.