Struct google_dataproc1::InstanceGroupConfig

pub struct InstanceGroupConfig {
    pub num_instances: Option<i32>,
    pub is_preemptible: Option<bool>,
    pub managed_group_config: Option<ManagedGroupConfig>,
    pub machine_type_uri: Option<String>,
    pub disk_config: Option<DiskConfig>,
    pub instance_names: Option<Vec<String>>,
    pub image_uri: Option<String>,
}

The config settings for Google Compute Engine resources in an instance group, such as a master or worker group.

This type is not used in any activity, and only used as part of another schema.

Fields

num_instances

The number of VM instances in the instance group. For master instance groups, must be set to 1.

is_preemptible

Specifies that this instance group contains Preemptible Instances.

managed_group_config

[Output-only] The config for Google Compute Engine Instance Group Manager that manages this group. This is only used for preemptible instance groups.

machine_type_uri

The Google Compute Engine machine type used for cluster instances. Example: https://www.googleapis.com/compute/v1/projects/[project_id]/zones/us-east1-a/machineTypes/n1-standard-2.

disk_config

Disk option config settings.

instance_names

The list of instance names. Cloud Dataproc derives the names from cluster_name, num_instances, and the instance group if not set by user (recommended practice is to let Cloud Dataproc derive the name).

image_uri

[Output-only] The Google Compute Engine image resource used for cluster instances. Inferred from SoftwareConfig.image_version.