Struct google_dataproc1::Job

pub struct Job {
    pub status: Option<JobStatus>,
    pub hadoop_job: Option<HadoopJob>,
    pub pyspark_job: Option<PySparkJob>,
    pub placement: Option<JobPlacement>,
    pub reference: Option<JobReference>,
    pub spark_sql_job: Option<SparkSqlJob>,
    pub pig_job: Option<PigJob>,
    pub driver_output_resource_uri: Option<String>,
    pub driver_control_files_uri: Option<String>,
    pub spark_job: Option<SparkJob>,
    pub status_history: Option<Vec<JobStatus>>,
    pub hive_job: Option<HiveJob>,
}

A Cloud Dataproc job resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

regions jobs submit projects (response)

regions jobs get projects (response)

regions jobs cancel projects (response)

Fields
status: Option<JobStatus>

[Output-only] The job status. Additional application-specific status information may be contained in the type_job and yarn_applications fields.

hadoop_job: Option<HadoopJob>

Job is a Hadoop job.

pyspark_job: Option<PySparkJob>

Job is a Pyspark job.

placement: Option<JobPlacement>

[Required] Job information, including how, when, and where to run the job.

reference: Option<JobReference>

[Optional] The fully qualified reference to the job, which can be used to obtain the equivalent REST path of the job resource. If this property is not specified when a job is created, the server generates a job_id.

spark_sql_job: Option<SparkSqlJob>

Job is a SparkSql job.

pig_job: Option<PigJob>

Job is a Pig job.

driver_output_resource_uri: Option<String>

[Output-only] A URI pointing to the location of the stdout of the job's driver program.

driver_control_files_uri: Option<String>

[Output-only] If present, the location of miscellaneous control files which may be used as part of job setup and handling. If not present, control files may be placed in the same location as driver_output_uri.

spark_job: Option<SparkJob>

Job is a Spark job.

status_history: Option<Vec<JobStatus>>

[Output-only] The previous job status.

hive_job: Option<HiveJob>

Job is a Hive job.