Struct google_dataproc1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Dataproc hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dataproc1 as dataproc1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dataproc1::Dataproc;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dataproc::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_autoscaling_policies_get_iam_policy(...)`, `locations_autoscaling_policies_set_iam_policy(...)`, `locations_autoscaling_policies_test_iam_permissions(...)`, `locations_workflow_templates_create(...)`, `locations_workflow_templates_delete(...)`, `locations_workflow_templates_get(...)`, `locations_workflow_templates_get_iam_policy(...)`, `locations_workflow_templates_instantiate(...)`, `locations_workflow_templates_instantiate_inline(...)`, `locations_workflow_templates_list(...)`, `locations_workflow_templates_set_iam_policy(...)`, `locations_workflow_templates_test_iam_permissions(...)`, `locations_workflow_templates_update(...)`, `regions_autoscaling_policies_get_iam_policy(...)`, `regions_autoscaling_policies_set_iam_policy(...)`, `regions_autoscaling_policies_test_iam_permissions(...)`, `regions_clusters_create(...)`, `regions_clusters_delete(...)`, `regions_clusters_diagnose(...)`, `regions_clusters_get(...)`, `regions_clusters_get_iam_policy(...)`, `regions_clusters_list(...)`, `regions_clusters_patch(...)`, `regions_clusters_set_iam_policy(...)`, `regions_clusters_test_iam_permissions(...)`, `regions_jobs_cancel(...)`, `regions_jobs_delete(...)`, `regions_jobs_get(...)`, `regions_jobs_get_iam_policy(...)`, `regions_jobs_list(...)`, `regions_jobs_patch(...)`, `regions_jobs_set_iam_policy(...)`, `regions_jobs_submit(...)`, `regions_jobs_test_iam_permissions(...)`, `regions_operations_cancel(...)`, `regions_operations_delete(...)`, `regions_operations_get(...)`, `regions_operations_get_iam_policy(...)`, `regions_operations_list(...)`, `regions_operations_set_iam_policy(...)`, `regions_operations_test_iam_permissions(...)`, `regions_workflow_templates_create(...)`, `regions_workflow_templates_delete(...)`, `regions_workflow_templates_get(...)`, `regions_workflow_templates_get_iam_policy(...)`, `regions_workflow_templates_instantiate(...)`, `regions_workflow_templates_instantiate_inline(...)`, `regions_workflow_templates_list(...)`, `regions_workflow_templates_set_iam_policy(...)`, `regions_workflow_templates_test_iam_permissions(...)` and `regions_workflow_templates_update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn regions_workflow_templates_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectRegionWorkflowTemplateTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_clusters_get(
&self,
project_id: &str,
region: &str,
cluster_name: &str
) -> ProjectRegionClusterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the resource representation for a cluster in a project.

Arguments

projectId - Required. The ID of the Google Cloud Platform project that the cluster belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

clusterName - Required. The cluster name.

pub fn regions_operations_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectRegionOperationTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_workflow_templates_list(
&self,
parent: &str
) -> ProjectLocationWorkflowTemplateListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists workflows that match the specified filter in the request.

Arguments

parent - Required. The "resource name" of the region, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}

pub fn regions_operations_get(
&self,
name: &str
) -> ProjectRegionOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn regions_workflow_templates_list(
&self,
parent: &str
) -> ProjectRegionWorkflowTemplateListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists workflows that match the specified filter in the request.

Arguments

parent - Required. The "resource name" of the region, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}

pub fn regions_operations_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectRegionOperationSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_workflow_templates_get(
&self,
name: &str
) -> ProjectLocationWorkflowTemplateGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the latest workflow template.Can retrieve previously instantiated template by specifying optional version parameter.

Arguments

name - Required. The "resource name" of the workflow template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn regions_jobs_patch(
&self,
request: Job,
project_id: &str,
region: &str,
job_id: &str
) -> ProjectRegionJobPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a job in a project.

Arguments

request - No description provided.

projectId - Required. The ID of the Google Cloud Platform project that the job belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

jobId - Required. The job ID.

pub fn regions_jobs_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectRegionJobTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_workflow_templates_delete(
&self,
name: &str
) -> ProjectRegionWorkflowTemplateDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a workflow template. It does not cancel in-progress workflows.

Arguments

name - Required. The "resource name" of the workflow template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn regions_workflow_templates_instantiate(
&self,
request: InstantiateWorkflowTemplateRequest,
name: &str
) -> ProjectRegionWorkflowTemplateInstantiateCall<'a, C, A>

Create a builder to help you perform the following task:

Instantiates a template and begins execution.The returned Operation can be used to track execution of workflow by polling operations.get. The Operation will complete when entire workflow is finished.The running workflow can be aborted via operations.cancel. This will cause any inflight jobs to be cancelled and workflow-owned clusters to be deleted.The Operation.metadata will be WorkflowMetadata.On successful completion, Operation.response will be Empty.

Arguments

request - No description provided.

name - Required. The "resource name" of the workflow template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn locations_workflow_templates_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationWorkflowTemplateTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_clusters_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectRegionClusterSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_workflow_templates_create(
&self,
request: WorkflowTemplate,
parent: &str
) -> ProjectLocationWorkflowTemplateCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates new workflow template.

Arguments

request - No description provided.

parent - Required. The "resource name" of the region, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}

pub fn regions_clusters_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectRegionClusterTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_clusters_patch(
&self,
request: Cluster,
project_id: &str,
region: &str,
cluster_name: &str
) -> ProjectRegionClusterPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a cluster in a project.

Arguments

request - No description provided.

projectId - Required. The ID of the Google Cloud Platform project the cluster belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

clusterName - Required. The cluster name.

pub fn regions_operations_delete(
&self,
name: &str
) -> ProjectRegionOperationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn locations_autoscaling_policies_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationAutoscalingPolicyGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_jobs_submit(
&self,
request: SubmitJobRequest,
project_id: &str,
region: &str
) -> ProjectRegionJobSubmitCall<'a, C, A>

Create a builder to help you perform the following task:

Submits a job to a cluster.

Arguments

request - No description provided.

projectId - Required. The ID of the Google Cloud Platform project that the job belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

pub fn regions_clusters_create(
&self,
request: Cluster,
project_id: &str,
region: &str
) -> ProjectRegionClusterCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a cluster in a project.

Arguments

request - No description provided.

projectId - Required. The ID of the Google Cloud Platform project that the cluster belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

pub fn locations_workflow_templates_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationWorkflowTemplateSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_workflow_templates_instantiate_inline(
&self,
request: WorkflowTemplate,
parent: &str
) -> ProjectLocationWorkflowTemplateInstantiateInlineCall<'a, C, A>

Create a builder to help you perform the following task:

Instantiates a template and begins execution.This method is equivalent to executing the sequence CreateWorkflowTemplate, InstantiateWorkflowTemplate, DeleteWorkflowTemplate.The returned Operation can be used to track execution of workflow by polling operations.get. The Operation will complete when entire workflow is finished.The running workflow can be aborted via operations.cancel. This will cause any inflight jobs to be cancelled and workflow-owned clusters to be deleted.The Operation.metadata will be WorkflowMetadata.On successful completion, Operation.response will be Empty.

Arguments

request - No description provided.

parent - Required. The "resource name" of the workflow template region, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}

pub fn locations_workflow_templates_update(
&self,
request: WorkflowTemplate,
name: &str
) -> ProjectLocationWorkflowTemplateUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates (replaces) workflow template. The updated template must contain version that matches the current server version.

Arguments

request - No description provided.

name - Output only. The "resource name" of the template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn regions_autoscaling_policies_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectRegionAutoscalingPolicySetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_workflow_templates_instantiate(
&self,
request: InstantiateWorkflowTemplateRequest,
name: &str
) -> ProjectLocationWorkflowTemplateInstantiateCall<'a, C, A>

Create a builder to help you perform the following task:

Instantiates a template and begins execution.The returned Operation can be used to track execution of workflow by polling operations.get. The Operation will complete when entire workflow is finished.The running workflow can be aborted via operations.cancel. This will cause any inflight jobs to be cancelled and workflow-owned clusters to be deleted.The Operation.metadata will be WorkflowMetadata.On successful completion, Operation.response will be Empty.

Arguments

request - No description provided.

name - Required. The "resource name" of the workflow template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn regions_clusters_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectRegionClusterGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_clusters_list(
&self,
project_id: &str,
region: &str
) -> ProjectRegionClusterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all regions/{region}/clusters in a project.

Arguments

projectId - Required. The ID of the Google Cloud Platform project that the cluster belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

pub fn regions_jobs_list(
&self,
project_id: &str,
region: &str
) -> ProjectRegionJobListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists regions/{region}/jobs in a project.

Arguments

projectId - Required. The ID of the Google Cloud Platform project that the job belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

pub fn regions_jobs_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectRegionJobGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_operations_cancel(
&self,
name: &str
) -> ProjectRegionOperationCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

name - The name of the operation resource to be cancelled.

pub fn regions_clusters_diagnose(
&self,
request: DiagnoseClusterRequest,
project_id: &str,
region: &str,
cluster_name: &str
) -> ProjectRegionClusterDiagnoseCall<'a, C, A>

Create a builder to help you perform the following task:

Gets cluster diagnostic information. After the operation completes, the Operation.response field contains DiagnoseClusterOutputLocation.

Arguments

request - No description provided.

projectId - Required. The ID of the Google Cloud Platform project that the cluster belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

clusterName - Required. The cluster name.

pub fn regions_workflow_templates_update(
&self,
request: WorkflowTemplate,
name: &str
) -> ProjectRegionWorkflowTemplateUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates (replaces) workflow template. The updated template must contain version that matches the current server version.

Arguments

request - No description provided.

name - Output only. The "resource name" of the template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn locations_autoscaling_policies_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationAutoscalingPolicySetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn regions_operations_list(
&self,
name: &str
) -> ProjectRegionOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users//operations. To override the binding, API services can add a binding such as "/v1/{name=users/}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_workflow_templates_delete(
&self,
name: &str
) -> ProjectLocationWorkflowTemplateDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a workflow template. It does not cancel in-progress workflows.

Arguments

name - Required. The "resource name" of the workflow template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn regions_workflow_templates_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectRegionWorkflowTemplateSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn regions_workflow_templates_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectRegionWorkflowTemplateGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_autoscaling_policies_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationAutoscalingPolicyTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_jobs_get(
&self,
project_id: &str,
region: &str,
job_id: &str
) -> ProjectRegionJobGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the resource representation for a job in a project.

Arguments

projectId - Required. The ID of the Google Cloud Platform project that the job belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

jobId - Required. The job ID.

pub fn locations_workflow_templates_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationWorkflowTemplateGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_clusters_delete(
&self,
project_id: &str,
region: &str,
cluster_name: &str
) -> ProjectRegionClusterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a cluster in a project.

Arguments

projectId - Required. The ID of the Google Cloud Platform project that the cluster belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

clusterName - Required. The cluster name.

pub fn regions_workflow_templates_instantiate_inline(
&self,
request: WorkflowTemplate,
parent: &str
) -> ProjectRegionWorkflowTemplateInstantiateInlineCall<'a, C, A>

Create a builder to help you perform the following task:

Instantiates a template and begins execution.This method is equivalent to executing the sequence CreateWorkflowTemplate, InstantiateWorkflowTemplate, DeleteWorkflowTemplate.The returned Operation can be used to track execution of workflow by polling operations.get. The Operation will complete when entire workflow is finished.The running workflow can be aborted via operations.cancel. This will cause any inflight jobs to be cancelled and workflow-owned clusters to be deleted.The Operation.metadata will be WorkflowMetadata.On successful completion, Operation.response will be Empty.

Arguments

request - No description provided.

parent - Required. The "resource name" of the workflow template region, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}

pub fn regions_autoscaling_policies_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectRegionAutoscalingPolicyTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_autoscaling_policies_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectRegionAutoscalingPolicyGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_operations_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectRegionOperationGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn regions_workflow_templates_get(
&self,
name: &str
) -> ProjectRegionWorkflowTemplateGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the latest workflow template.Can retrieve previously instantiated template by specifying optional version parameter.

Arguments

name - Required. The "resource name" of the workflow template, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}/workflowTemplates/{template_id}

pub fn regions_workflow_templates_create(
&self,
request: WorkflowTemplate,
parent: &str
) -> ProjectRegionWorkflowTemplateCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates new workflow template.

Arguments

request - No description provided.

parent - Required. The "resource name" of the region, as described in https://cloud.google.com/apis/design/resource_names of the form projects/{project_id}/regions/{region}

pub fn regions_jobs_delete(
&self,
project_id: &str,
region: &str,
job_id: &str
) -> ProjectRegionJobDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the job from the project. If the job is active, the delete fails, and the response returns FAILED_PRECONDITION.

Arguments

projectId - Required. The ID of the Google Cloud Platform project that the job belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

jobId - Required. The job ID.

pub fn regions_jobs_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectRegionJobSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn regions_jobs_cancel(
&self,
request: CancelJobRequest,
project_id: &str,
region: &str,
job_id: &str
) -> ProjectRegionJobCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts a job cancellation request. To access the job resource after cancellation, call regions/{region}/jobs.list or regions/{region}/jobs.get.

Arguments

request - No description provided.

projectId - Required. The ID of the Google Cloud Platform project that the job belongs to.

region - Required. The Cloud Dataproc region in which to handle the request.

jobId - Required. The job ID.