Struct google_dataproc1::SoftwareConfig

pub struct SoftwareConfig {
    pub image_version: Option<String>,
    pub properties: Option<HashMap<String, String>>,
}

Specifies the selection and config of software inside the cluster.

This type is not used in any activity, and only used as part of another schema.

Fields

image_version

[Optional] The version of software inside the cluster. It must match the regular expression [0-9]+\.[0-9]+. If unspecified, it defaults to the latest version (see Cloud Dataproc Versioning).

properties

[Optional] The properties to set on daemon config files. Property keys are specified in prefix:property format, such as core:fs.defaultFS. The following are supported prefixes and their mappings: * core: core-site.xml * hdfs: hdfs-site.xml * mapred: mapred-site.xml * yarn: yarn-site.xml * hive: hive-site.xml * pig: pig.properties * spark: spark-defaults.conf